Move ref validation to Ref class
